requirements.txt
fastapi
uvicorn
redis
pika
ultralytics==8.1.34
opencv-python>=4.8.0
numpy<2.0
PyYAML>=6.0
easyocr>=1.7.1
torch==2.1.2
torchvision==0.16.2
tqdm
scikit-learn
python-multipart